ref(strings): moved to the appropriate folder
